+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/mkv/MatroskaExtractor.java
@@ -54,8 +54,10 @@ import java.nio.ByteOrder;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.UUID;
 import org.checkerframework.checker.nullness.compatqual.NullableType;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
@@ -318,6 +320,18 @@ public class MatroskaExtractor implements Extractor {
    * Sub format for PCM.
    */
   private static final UUID WAVE_SUBFORMAT_PCM = new UUID(0x0100000000001000L, 0x800000AA00389B71L);
+
+  /** Some HTC devices signal rotation in track names. */
+  private static final Map<String, Integer> TRACK_NAME_TO_ROTATION_DEGREES;
+
+  static {
+    Map<String, Integer> trackNameToRotationDegrees = new HashMap<>();
+    trackNameToRotationDegrees.put("htc_video_rotA-000", 0);
+    trackNameToRotationDegrees.put("htc_video_rotA-090", 90);
+    trackNameToRotationDegrees.put("htc_video_rotA-180", 180);
+    trackNameToRotationDegrees.put("htc_video_rotA-270", 270);
+    TRACK_NAME_TO_ROTATION_DEGREES = Collections.unmodifiableMap(trackNameToRotationDegrees);
+  }
 
   private final EbmlReader reader;
   private final VarintReader varintReader;
@@ -2088,15 +2102,9 @@ public class MatroskaExtractor implements Extractor {
           colorInfo = new ColorInfo(colorSpace, colorRange, colorTransfer, hdrStaticInfo);
         }
         int rotationDegrees = Format.NO_VALUE;
-        // Some HTC devices signal rotation in track names.
-        if ("htc_video_rotA-000".equals(name)) {
-          rotationDegrees = 0;
-        } else if ("htc_video_rotA-090".equals(name)) {
-          rotationDegrees = 90;
-        } else if ("htc_video_rotA-180".equals(name)) {
-          rotationDegrees = 180;
-        } else if ("htc_video_rotA-270".equals(name)) {
-          rotationDegrees = 270;
+
+        if (TRACK_NAME_TO_ROTATION_DEGREES.containsKey(name)) {
+          rotationDegrees = TRACK_NAME_TO_ROTATION_DEGREES.get(name);
         }
         if (projectionType == C.PROJECTION_RECTANGULAR
             && Float.compare(projectionPoseYaw, 0f) == 0
@@ -2134,6 +2142,10 @@ public class MatroskaExtractor implements Extractor {
         type = C.TRACK_TYPE_TEXT;
       } else {
         throw new ParserException("Unexpected MIME type.");
+      }
+
+      if (!TRACK_NAME_TO_ROTATION_DEGREES.containsKey(name)) {
+        formatBuilder.setLabel(name);
       }
 
       Format format =